♻️ Refact: 기존 NetworkService 프로토콜을 Protocol/Service 폴더에 파일 생성 후 이동
