+++ EatPic-iOS/Sources/Common/Protocol/TokenProviding.swift
