+++ EatPic-iOS/Sources/Core/Networks/NetworkService.swift
@@ -9,24 +9,6 @@ import Foundation
 import Moya
 import Alamofire
 
-protocol NetworkService {
-    var isTokenExpiringSoon: Bool { get }
-    
-    func createProvider<T: TargetType>(
-        for targetType: T.Type,
-        additionalPlugins: [PluginType]
-    ) -> MoyaProvider<T>
-    
-    func testProvider<T: TargetType>(for targetType: T.Type) -> MoyaProvider<T>
-}
-
-/// 기본값을 제공하는 extension
-extension NetworkService {
-    func createProvider<T: TargetType>(for targetType: T.Type) -> MoyaProvider<T> {
-        return createProvider(for: targetType, additionalPlugins: [])
-    }
-}
-
 class NetworkServiceImpl: @unchecked Sendable, NetworkService {
     private let tokenProvider: TokenProviding
     private let accessTokenRefresher: AccessTokenRefresher
